+++ greenfield/eldenring/data.py
@@ -5560,4 +5560,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:3ef25de73614ade8e9717736838426b00a98043d106849656a331ec4eafa718b', 'module': 'data.py', 'body_sha256': 'sha256:ce9920d78f81dc16c4cee592a671cad8a7b247bd6101309f4aa2399562f30a8b'}
+_GEN_STAMP = {'inputs_hash': 'sha256:7d7ea51db2b0c64df527976e873f77ff50a785f3bc640f25b05aaeb2c87113e6', 'module': 'data.py', 'body_sha256': 'sha256:ce9920d78f81dc16c4cee592a671cad8a7b247bd6101309f4aa2399562f30a8b'}